+++ realm/realm-library/src/objectServer/java/io/realm/SyncSession.java
@@ -159,9 +159,10 @@ public class SyncSession {
         ErrorCode errCode = ErrorCode.fromInt(errorCode);
         if (errCode == ErrorCode.CLIENT_RESET) {
             // errorMessage contains the path to the backed up file
+            SyncConfiguration offlineSyncConfiguration = SyncConfiguration.forOfflineSync(configuration.getSchemaMediator(), errorMessage);
             errorHandler.onError(this, new ClientResetRequiredError(errCode, "A Client Reset is required. " +
                     "Read more here: https://realm.io/docs/realm-object-server/#client-recovery-from-a-backup.",
-                    errorMessage, getConfiguration()));
+                    configuration, offlineSyncConfiguration));
         } else {
             errorHandler.onError(this, new ObjectServerError(errCode, errorMessage));
         }